requirements.txt
# Core
alpaca-trade-api==3.0.2
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
schedule==1.2.1
pytz==2024.1

# Technical indicators
pandas-ta==0.3.14b0

# Telegram alerts
python-telegram-bot==20.7

# Dashboard / logging
rich==13.7.0
tabulate==0.9.0

# NSE India data
nsepython==1.11

# Async support
aiohttp==3.9.3
